Update of the Contact Me page
This is an update to the Contact Me page.

Features removed:
WhatsApp integration and QR image
Corresponding buttons to share via whatsapp.

Features added:
Contact us and address with information.
Textareas for name, email, and message that wishes to be sent.
Buttons to send clear the message.

Notes:
No functionality on buttons and textareas added.
Hover animations/ focus animations not implemented.